+++ test/OpenTelemetry.Instrumentation.Http.Tests/HttpClientTests.Basic.netcore31.cs
@@ -21,6 +21,7 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using Moq;
+using OpenTelemetry.Context;
 using OpenTelemetry.Context.Propagation;
 using OpenTelemetry.Instrumentation.Http.Implementation;
 using OpenTelemetry.Tests;
@@ -248,6 +249,43 @@ namespace OpenTelemetry.Instrumentation.Http.Tests
             }
 
             Assert.Equal(2, processor.Invocations.Count); // OnShutdown/Dispose called.
+        }
+
+        [Fact]
+        public async Task HttpClientInstrumentationCorrelationAndBaggage()
+        {
+            var activityProcessor = new Mock<ActivityProcessor>();
+
+            using var parent = new Activity("w3c activity");
+            parent.SetIdFormat(ActivityIdFormat.W3C);
+            parent.AddBaggage("k1", "v1");
+            parent.ActivityTraceFlags = ActivityTraceFlags.Recorded;
+            parent.Start();
+
+            Baggage.SetBaggage("k2", "v2");
+
+            using (Sdk.CreateTracerProviderBuilder()
+                .AddHttpClientInstrumentation()
+                .AddProcessor(activityProcessor.Object)
+                .Build())
+            {
+                using var c = new HttpClient();
+                using var r = await c.GetAsync("https://opentelemetry.io/").ConfigureAwait(false);
+            }
+
+            Assert.Equal(4, activityProcessor.Invocations.Count);
+
+            var activity = (Activity)activityProcessor.Invocations[1].Arguments[0];
+
+            HttpRequestMessage thisRequest = (HttpRequestMessage)activity.GetCustomProperty(HttpHandlerDiagnosticListener.RequestCustomPropertyName);
+
+            string[] correlationContext = thisRequest.Headers.GetValues("Correlation-Context").First().Split(',');
+            Assert.Single(correlationContext);
+            Assert.Contains("k1=v1", correlationContext);
+
+            string[] baggage = thisRequest.Headers.GetValues("Baggage").First().Split(',');
+            Assert.Single(baggage);
+            Assert.Contains("k2=v2", baggage);
         }
 
         public void Dispose()